Rename css completion module name
We use lower case, no underscore and relatively short module names.
